Exceptions in ReservationDAOImpl are now caught and CarDAOTest class is done

+++ vozovyPark/src/main/java/cz/muni/fi/pa165/vozovypark/DAO/ReservationDAOImpl.java
@@ -26,6 +26,9 @@ public class ReservationDAOImpl implements ReservationDAO {
     }
    
     public Reservation getReservationById(Long id){
+        if(id==null) {
+            throw new IllegalArgumentException("ID cannot be null");
+        }
         EntityManager entityManager = this.entityManagerFactory.createEntityManager();
         Query q = entityManager.createNamedQuery(Reservation.FIND_BY_ID);
         q.setParameter("id", id);
@@ -33,21 +36,42 @@ public class ReservationDAOImpl implements ReservationDAO {
     }
     
     public void insert(Reservation r) {
+        if(r == null) {
+            throw new IllegalArgumentException("You must specify reservation to insert");
+        }
         EntityManager entityManager = this.entityManagerFactory.createEntityManager();
         entityManager.persist(r);
     }
     
     public void update(Reservation r) {
+        if(r==null){
+            throw new IllegalArgumentException("You must specify reservation to update");
+        }
+        if(r.getId()==null) {
+            throw new IllegalArgumentException("Reservation to update must have specified ID");
+        }
         EntityManager entityManager = this.entityManagerFactory.createEntityManager();
         entityManager.merge(r);
     }
     
     public void remove(Reservation r) {
+        if(r==null){
+            throw new IllegalArgumentException("You must specify reservation to remove");
+        }
+        if(r.getId()==null) {
+            throw new IllegalArgumentException("Reservation to remove must have specified ID");
+        }
         EntityManager entityManager = this.entityManagerFactory.createEntityManager();
         entityManager.remove(entityManager.merge(r));
     }
 
   public List<Reservation> getReservationByCar(Car car) {
+      if(car==null){
+            throw new IllegalArgumentException("You must specify reservation car");
+        }
+      if(car.getId()==null) {
+            throw new IllegalArgumentException("Car must have specified ID");
+        }
       EntityManager entityManager = this.entityManagerFactory.createEntityManager();
        Query q = entityManager.createNamedQuery(Reservation.FIND_BY_CAR);
        q.setParameter("car", car);
@@ -55,6 +79,12 @@ public class ReservationDAOImpl implements ReservationDAO {
    }
    
    public List<Reservation> getReservationByEmployee(Employee employee) {
+       if(employee==null){
+            throw new IllegalArgumentException("You must specify reservation employee");
+        }
+       if(employee.getId()==null) {
+            throw new IllegalArgumentException("Employee must have specified ID");
+        }
        EntityManager entityManager = this.entityManagerFactory.createEntityManager();
        Query q = entityManager.createNamedQuery(Reservation.FIND_BY_EMPLOYEE);
        q.setParameter("employee", employee);
@@ -62,6 +92,18 @@ public class ReservationDAOImpl implements ReservationDAO {
    }
    
    public List<Reservation> getReservationByCarAndEmployee(Car car, Employee employee) {
+       if(car==null){
+            throw new IllegalArgumentException("You must specify reservation car");
+        }
+      if(car.getId()==null) {
+            throw new IllegalArgumentException("Car must have specified ID");
+        }
+      if(employee==null){
+            throw new IllegalArgumentException("You must specify reservation employee");
+        }
+       if(employee.getId()==null) {
+            throw new IllegalArgumentException("Employee must have specified ID");
+        }
        EntityManager entityManager = this.entityManagerFactory.createEntityManager();
        Query q = entityManager.createNamedQuery(Reservation.FIND_BY_CAR_AND_EMPLOYEE);
        q.setParameter("car", car);
